Fix CesiumIonSession race condition.
Sometimes, during the tests, the managed CesiumIonSession could get
garbage collected before the requests made as part of the "resume"
completed. Because we were using the corresponding Impl class in the
continuations, and the Impl class's lifetime is controlled by the
lifetime of the managed object, this would sometimes cause a crash. The
solution implemented here is to make sure the continuations also keep
the managed instance alive. We also check to see if the held managed
instance becomes nullptr, which handles causes like AppDomain unload.