Attempt to work around warning spam from Unity physics engines.
"Detected one or more triangles where the distance between any 2
vertices is greater than 500 units. The resulting Triangle Mesh can
impact simulation and query stability. It is recommended to tessellate
meshes that have large triangles."

This commit tries to avoid this warning by scaling down the vertex
positions of every tile mesh to fit inside a cube with a diagonal <= 1
unit. And then scale it back up to its proper size via the
transformation matrix.

It reduces the number of occurrences of this warning, but for some
reason it doesn't eliminate them entirely.