Fixed cross compilation from Linux
Example: build happens from Linux CI server, with another platform (like WebGL) as target. CesiumForUnity package is installed.

The build happens by from the Unity Editor on Linux. When the editor is opened, it does two things:

* loads the project and ensures that the editor loads the needed resources/scripts, compiles the needed scripts
* switches to the target platform to make the build, starts the build, etc

Cesium does not support building on Linux: the native runtime is missing for Linux, and the generated code for from Reinterop does not generate any code for Linux.
However, when loading the Unity Editor on Linux, it tries to compile Cesium scripts, which fail because of missing Reinterop bindings. When that fails, the cross compilation to WebGL does not start at all.

Cesium .asmdef files don't declare explicit support for Linux, which is ok. However, they declare support for "Editor", which also matches the Editor on Linux. With this commit, we ensure that Cesium scripts won't be loaded for the Editor on Linux either. This way, projects can be loaded on Linux and builds for a different target platform successfully.