Attempt to use AppDomain `Id` to check valid ObjectHandle usage.
Unfortunately, it doesn't work. The problem is that Mono pretty
aggressively reuses AppDomain IDs. So even if an ObjectHandle was
created in an AppDomain with a particular ID, and later we find that
current AppDomain has the same ID, it doesn't conclusively prove that
the AppDomain hasn't changed. `GetHashCode` is not reliable,
eitheAttempt to use AppDomain `Id` to check valid ObjectHandle usage.

Unfortunately, it doesn't work. The problem is that Mono pretty
aggressively reuses AppDomain IDs. So even if an ObjectHandle was
created in an AppDomain with a particular ID, and later we find that
current AppDomain has the same ID, it doesn't conclusively prove that
the AppDomain hasn't changed. `GetHashCode` is not reliable, either.